1008-Third commit：Modified for paddle and solved the problem of tense.permute() function with transpose:with different parameters. New problem: for the 61 lines of attn.py, I paddle in a different way from torch, so I don't know how to apply them.-第三次提交：针对paddle进行了修改，解决了tensor.permute()函数的对应问题，使用transpose(参数不一样的)进行替换。 新问题：attn.py中61行,paddle的tensor切片方式与torch不一样，还不知道如何对应。

+++ exp/exp_informer.py
@@ -157,7 +157,8 @@ class Exp_Informer(Exp_Basic):
                 
                 model_optim.clear_grad()  # mf-梯度清零，防止梯度累加
                 pred, true = self._process_one_batch(
-                    train_data, batch_x, batch_y, batch_x_mark, batch_y_mark)
+                    train_data, batch_x.astype('float32'), batch_y.astype('float32'),
+                    batch_x_mark.astype('float32'), batch_y_mark.astype('float32'))
                 loss = criterion(pred, true)
                 train_loss.append(loss.item())
                 
@@ -270,7 +271,7 @@ class Exp_Informer(Exp_Basic):
 
         # decoder input
         if self.args.padding==0:
-            dec_inp = paddle.zeros([batch_y.shape[0], self.args.pred_len, batch_y.shape[-1]], dtype='float64')
+            dec_inp = paddle.zeros([batch_y.shape[0], self.args.pred_len, batch_y.shape[-1]], dtype='float32')
         elif self.args.padding==1:
             dec_inp = paddle.ones([batch_y.shape[0], self.args.pred_len, batch_y.shape[-1]])
         para = batch_y[:,:self.args.label_len,:]
